Moving to modular Sinatra
