+++ config.ru
@@ -1,4 +1,4 @@
 $stdout.sync = true
 
-require './whedon'
-run Sinatra::Application
+require './whedon_api'
+run WhedonApi
