+++ spec/spec_helper.rb
@@ -5,12 +5,16 @@ WebMock.disable_net_connect!(allow_localhost: true)
 
 ENV['RACK_ENV'] = 'test'
 
-require File.expand_path '../../whedon.rb', __FILE__
+require File.expand_path '../../whedon_api.rb', __FILE__
 require File.expand_path '../support/fake_github.rb', __FILE__
 
 module RSpecMixin
   include Rack::Test::Methods
-  def app() Sinatra::Application end
+  def app() described_class end
+
+  def json_fixture(file_name)
+    File.open(File.dirname(__FILE__) + '/support/fixtures/' + file_name, 'rb').read
+  end
 end
 
 RSpec.configure do |config|
