+++ spec/support/fixtures/junk-payload.json
@@ -1,0 +1,4 @@
+{
+  "i": "am",
+  "not": "valid"
+}
